fix(discord): casse l'import circulaire qui plantait le boot
La consolidation en un seul command-registry.ts créait un cycle: les modules
de commandes (change-captain, intercept-onboarding) importaient des helpers
depuis command-registry.ts, qui lui-même importe tous les tableaux de commandes.
Au boot, interactionRouter charge crew/index en premier -> change-captain ->
command-registry, dont le corps lit crewCommands avant que crew/index ait fini
de s'initialiser (TDZ: 'Cannot access crewCommands before initialization').

Sépare les helpers purs (getCommandDisplayName, getCommandKeywords) dans
discord/command-names.ts (ne dépend que de types.js, donc sans cycle).
command-registry.ts garde le registre eager + resolveCommand et n'est plus
importé que par router.ts.
